+++ build.gradle
@@ -69,6 +69,9 @@ dependencies {
 
     // caffeine cache
     implementation 'com.github.ben-manes.caffeine:caffeine'
+
+    // kafka
+    implementation 'org.springframework.kafka:spring-kafka'
 }
 
 tasks.named('test') {
